firstshowcss

+++ db/schema.rb
@@ -12,7 +12,6 @@
 
 ActiveRecord::Schema.define(version: 2022_05_25_123050) do
 
-
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -45,8 +44,8 @@ ActiveRecord::Schema.define(version: 2022_05_25_123050) do
   end
 
   create_table "bookings", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "superpower_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "superpower_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.date "starts_at"
@@ -67,7 +66,7 @@ ActiveRecord::Schema.define(version: 2022_05_25_123050) do
     t.string "name"
     t.text "description"
     t.integer "price"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["user_id"], name: "index_superpowers_on_user_id"
